+++ NEWS.md
@@ -1,3 +1,8 @@
+# saqgetr 0.1.18
+
+  - Explicitly open and close remote connections to avoid the use of `closeAllConnections` in `get_saq_observations`. 
+  - Make messaging simpler. 
+
 # saqgetr 0.1.17
 
   - Resubmisison after CRAN feedback.